macvlans: Make vip accesible using GORB
Register docker service as a GORB service so that the docker service is
routed via vip.

Usage:

    1. Start a gorb router.

        ./gorb &

    2. Create a macvlans network using the gorb.

        docker network create macvlans1 \
            -d macvlans \
            --subnet=10.113.136.0/22 \
            --ip-range=10.113.136.32/27 \
            --gateway=10.113.136.1 \
            -o parent=bond0 \
            --label=gorb.url=http://10.113.129.9:4672

    3. Create a service, using the network, whose ports are published.

        docker service create --name=web1 --network=macvlans1 --publish 80:80

Note 1: Published port, which is specified by `--publish` option, must
be same with the target port because of limitation of ip tunneling.

Note 2: All packets to the service vip should go to the GORB first. It
means that the machine running GORB should have an interface which
contains the vip address. Unfortunately this commit can't do that.